// resources/js/admin/locations/types/location-form-errors.ts
export default interface LocationFormErrors {
  name?: string;
  description?: string;
  quest_reward_item_id?: string;
  required_quest_item_id?: string;
  x?: string;
  y?: string;
  type?: string;
  pin_css_class?: string;
  hours_to_drop?: string;
  minutes_between_delve_fights?: string;
  form?: string;
}